用户登录，service和Controller开发

+++ src/main/java/io/whysff/community/service/UserService.java
@@ -13,5 +13,5 @@ public interface UserService {
 
     int activation(int userId, String code);
 
-    Map<String, Object> login(String username, String password, int expiredSeconds);
+    Map<String, Object> login(String username, String password, long expiredSeconds);
 }
